requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==8.3.1
requests==2.32.3
selenium==4.24.0
webdriver-manager==4.0.2